+++ examples/_scheduled/start-world.job.ts
@@ -1,4 +1,4 @@
-import { defineJob } from '@core/mod.ts';
+import { defineJob } from '@hushkey/hound/mod.ts';
 
 export const startWorldJob = defineJob<{ foo: string }, { email: string }>(
   'start-world',
@@ -10,11 +10,11 @@ export const startWorldJob = defineJob<{ foo: string }, { email: string }>(
       'color: white; background-color: red;',
     );
 
-    for (let i = 0; i < 100; i++) {
-      ctx.emit('user.read', {
-        email: 'john.doe@example.com',
-      }, { id: `on-request-${i + 1}`, queue: 'tasks' });
-    }
+    // for (let i = 0; i < 10_000; i++) {
+    //   ctx.emit('user.read', {
+    //     email: 'john.doe@example.com',
+    //   }, { id: `on-request-${i + 1}`, queue: 'tasks' });
+    // }
   },
   {
     repeat: {